+++ TerrainSDK/minidata/jpegbase.h
@@ -6,7 +6,7 @@
 namespace jpegbase {
 
 unsigned char *decompressJPEGimage(unsigned char *data,unsigned int bytes,int *width,int *height,int *components);
-void compressJPEGimage(unsigned char *image,int width,int height,int components,float quality,unsigned char **data,unsigned int *bytes);
+void compressJPEGimage(unsigned char *image,int width,int height,int components,float quality,unsigned char **data,long long *bytes);
 
 }
